refactor(test): clean up post-Jest migration test infrastructure
Remove one-time Jest migration scripts and @types/jest, fix stale playbook
docs, split web.preload into setup modules, consolidate test launchers via
runner-utils, and add mongo-free :fast paths for backend/sync/scripts using
synthetic COMPASS_TEST_MONGO_URI env (no mongod boot).
