+++ package.json
@@ -33,13 +33,14 @@
     "test:a11y": "bunx playwright test e2e/accessibility",
     "test:e2e": "bunx playwright test",
     "test:backend": "bun packages/scripts/src/testing/test-mongo-env.ts backend --",
-    "test:backend:fast": "bun packages/scripts/src/testing/test-mongo-env.ts backend -- --path-ignore-patterns '**/*.db.test.ts'",
+    "test:backend:fast": "bun packages/scripts/src/testing/test-parallel.ts backend-fast -- --path-ignore-patterns '**/*.db.test.ts'",
     "test:backend:db": "bun packages/scripts/src/testing/test-mongo-env.ts backend -- './packages/backend/src/**/*.db.test.ts'",
-    "test:core": "TZ=Etc/UTC bun test --parallel --preload ./packages/scripts/src/testing/core.preload.ts ./packages/core/src",
+    "test:core": "bun packages/scripts/src/testing/test-parallel.ts core --",
     "test:sync": "bun packages/scripts/src/testing/test-mongo-env.ts sync --",
+    "test:sync:fast": "bun packages/scripts/src/testing/test-parallel.ts sync-fast -- --path-ignore-patterns '**/*.db.test.ts'",
     "test:web": "bun packages/scripts/src/testing/test-isolated.ts",
     "test:scripts": "bun packages/scripts/src/testing/test-mongo-env.ts scripts --",
-    "test:scripts:fast": "bun packages/scripts/src/testing/test-mongo-env.ts scripts -- --path-ignore-patterns '**/*.db.test.ts'",
+    "test:scripts:fast": "bun packages/scripts/src/testing/test-parallel.ts scripts-fast -- --path-ignore-patterns '**/*.db.test.ts'",
     "test:scripts:db": "bun packages/scripts/src/testing/test-mongo-env.ts scripts -- './packages/scripts/src/**/*.db.test.ts'",
     "test:migrations": "bun packages/scripts/src/testing/test-mongo-env.ts scripts -- './packages/scripts/src/migrations/**/*.db.test.ts'",
     "type-check": "bunx typescript@7.0.2 --noEmit && bunx typescript@7.0.2 -p packages/web/tsconfig.app.json --noEmit && bun run type-check:web-tests",
