+++ packages/backend/package.json
@@ -28,7 +28,6 @@
     "@faker-js/faker": "^9.6.0",
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.13",
-    "@types/jest": "^29.0.0",
     "@types/lodash.mergewith": "^4.6.9",
     "@types/node": "^22.13.10",
     "@types/qs": "^6.9.18",
