+++ packages/scripts/src/testing/scripts.preload.fast.ts
@@ -1,0 +1,2 @@
+// Fast tier for scripts tests: same backend env/seams as backend-fast, no mongod.
+import "@backend/__tests__/backend.preload.fast";
