+++ packages/scripts/src/testing/test-isolated.ts
@@ -2,27 +2,16 @@
  * Runs web tests with one `bun test` process per file so preload mock.module
  * state is not cleared by Bun's --isolate (which --parallel enables).
  */
-import { Glob } from "bun";
+
+import { formatSummary, resolveTestFiles, runPool } from "./runner-utils";
 import { cpus } from "node:os";
 import { resolve } from "node:path";
 
 const PRELOAD = resolve("packages/web/src/__tests__/web.preload.ts");
 const SCAN = "packages/web/src/**/*.{test,spec}.{ts,tsx}";
 
-function resolveFiles(extraArgs: string[]): string[] {
-  const explicit = extraArgs.filter((arg) => !arg.startsWith("-"));
-  if (explicit.length > 0) {
-    return explicit.map((arg) => (arg.startsWith("./") ? arg : `./${arg}`));
-  }
-
-  return Array.from(new Glob(SCAN).scanSync("."))
-    .map((file) => `./${file.replace(/^\.\//, "")}`)
-    .sort();
-}
-
 const extraArgs = process.argv.slice(2);
-const files = resolveFiles(extraArgs);
-const bunFlags = extraArgs.filter((arg) => arg.startsWith("-"));
+const { files, bunFlags } = resolveTestFiles(SCAN, extraArgs);
 
 if (files.length === 0) {
   console.error("No test files found");
@@ -30,7 +19,6 @@ if (files.length === 0) {
 }
 
 const concurrency = Math.max(1, Math.min(cpus().length, 8));
-let nextIndex = 0;
 let passedFiles = 0;
 const failedFiles: string[] = [];
 
@@ -68,22 +56,12 @@ async function runOne(index: number): Promise<void> {
   }
 }
 
-async function worker(): Promise<void> {
-  while (nextIndex < files.length) {
-    const index = nextIndex++;
-    await runOne(index);
-  }
-}
-
 const started = Date.now();
 console.log(`Running ${files.length} web test files...`);
 
-await Promise.all(Array.from({ length: concurrency }, () => worker()));
+await runPool(concurrency, files.length, runOne);
 
-const seconds = ((Date.now() - started) / 1000).toFixed(1);
-console.log(
-  `\n\nweb: ${passedFiles}/${files.length} files passed | ${seconds}s`,
-);
+console.log(`\n\n${formatSummary("web", passedFiles, files.length, started)}`);
 
 if (failedFiles.length > 0) {
   console.log(`Failed files:\n  ${failedFiles.join("\n  ")}`);
