+++ packages/scripts/src/testing/test-mongo-env.ts
@@ -8,9 +8,9 @@
  * Usage:
  *   bun test-mongo-env.ts <backend|scripts|sync> -- [bun test flags/paths...]
  */
-import { Glob } from "bun";
 import { MongoMemoryReplSet } from "mongodb-memory-server";
 import { backendTestSpawnEnv } from "./backend-test-env";
+import { formatDuration, resolveTestTargets } from "./runner-utils";
 import { resolve } from "node:path";
 
 type PackageName = "backend" | "scripts" | "sync";
@@ -36,73 +36,6 @@ const PACKAGES: Record<
   },
 };
 
-function parseExtraArgs(extraArgs: string[]): {
-  bunFlags: string[];
-  ignorePattern?: string;
-  explicitPaths: string[];
-} {
-  const bunFlags: string[] = [];
-  const explicitPaths: string[] = [];
-  let ignorePattern: string | undefined;
-
-  for (let i = 0; i < extraArgs.length; i++) {
-    const arg = extraArgs[i]!;
-
-    if (arg === "--path-ignore-patterns") {
-      ignorePattern = extraArgs[i + 1];
-      i++;
-      continue;
-    }
-
-    if (arg.startsWith("-")) {
-      bunFlags.push(arg);
-      continue;
-    }
-
-    if (!arg.includes("*")) {
-      explicitPaths.push(arg.startsWith("./") ? arg : `./${arg}`);
-      continue;
-    }
-
-    const matches = Array.from(new Glob(arg).scanSync("."));
-    if (matches.length === 0) {
-      console.error(`No test files matched: ${arg}`);
-      process.exit(1);
-    }
-
-    for (const match of matches) {
-      explicitPaths.push(
-        match.startsWith("./") ? match : `./${match.replace(/^\.\//, "")}`,
-      );
-    }
-  }
-
-  return { bunFlags, ignorePattern, explicitPaths };
-}
-
-function resolveTestTargets(
-  pkg: PackageName,
-  scan: string,
-  extraArgs: string[],
-): { targets: string[]; bunFlags: string[]; label: string } {
-  const { bunFlags, ignorePattern, explicitPaths } = parseExtraArgs(extraArgs);
-
-  if (explicitPaths.length > 0) {
-    return {
-      targets: explicitPaths,
-      bunFlags,
-      label: `${explicitPaths.length} ${pkg} test files`,
-    };
-  }
-
-  const flags = [...bunFlags];
-  if (ignorePattern) {
-    flags.push("--path-ignore-patterns", ignorePattern);
-  }
-
-  return { targets: [scan], bunFlags: flags, label: scan };
-}
-
 const pkg = process.argv[2] as PackageName;
 const separatorIndex = process.argv.indexOf("--");
 const extraArgs =
@@ -119,7 +52,7 @@ if (!pkg || !PACKAGES[pkg]) {
 
 const { preload, scan } = PACKAGES[pkg];
 const preloadPath = resolve(preload);
-const { targets, bunFlags, label } = resolveTestTargets(pkg, scan, extraArgs);
+const { targets, bunFlags, label } = resolveTestTargets(scan, extraArgs);
 
 const started = Date.now();
 
@@ -152,8 +85,7 @@ try {
     stderr: "inherit",
   });
   const code = await proc.exited;
-  const seconds = ((Date.now() - started) / 1000).toFixed(1);
-  console.log(`\n${pkg}: finished in ${seconds}s`);
+  console.log(`\n${pkg}: finished in ${formatDuration(started)}s`);
 
   if (code !== 0) {
     process.exit(code ?? 1);
