+++ packages/sync/src/__tests__/sync.preload.fast.ts
@@ -1,0 +1,3 @@
+// Fast tier: env only. Non-db sync tests never connect to Mongo.
+process.env["NODE_ENV"] = "test";
+process.env["LOG_LEVEL"] = "debug";
