+++ packages/web/src/__tests__/web.preload.ts
@@ -1,356 +1,18 @@
 // sort-imports-ignore — side-effect import order matters
 
-import { JSDOM } from "jsdom";
-import {
-  afterAll,
-  afterEach,
-  beforeAll,
-  beforeEach,
-  expect,
-  mock,
-} from "bun:test";
-import { createRequire } from "node:module";
-
 import "@core/__tests__/core.test.init";
 import "@core/__tests__/core.test.start";
 import "./web.test.init";
+
+import "./setup/asset-stubs.plugin";
+import "./setup/jsdom-env";
+import "./setup/browser-polyfills";
+import "./setup/indexeddb-env";
+
 import { mockNodeModules } from "./__mocks__/mock.setup";
-import { server } from "./__mocks__/server/mock.server";
-
-const requireMatchers = createRequire(import.meta.path);
-const jestDomMatchers = requireMatchers(
-  "@testing-library/jest-dom/dist/matchers.js",
-) as Record<string, unknown>;
-expect.extend(jestDomMatchers);
-(globalThis as typeof globalThis & { expect: typeof expect }).expect = expect;
-
-Bun.plugin({
-  name: "web-test-asset-stubs",
-  setup(build) {
-    build.onResolve({ filter: /\.(css|less)$/ }, () => ({
-      path: "virtual:web-test-empty-style",
-      namespace: "web-test-stub",
-    }));
-
-    build.onResolve({ filter: /\.(jpe?g|png|gif)$/i }, () => ({
-      path: "virtual:web-test-file-stub",
-      namespace: "web-test-stub",
-    }));
-
-    build.onResolve({ filter: /\.svg$/ }, () => ({
-      path: "virtual:web-test-svg-stub",
-      namespace: "web-test-stub",
-    }));
-
-    build.onLoad({ filter: /.*/, namespace: "web-test-stub" }, (args) => {
-      if (args.path === "virtual:web-test-empty-style") {
-        return { contents: "export default {};", loader: "js" };
-      }
-
-      if (args.path === "virtual:web-test-file-stub") {
-        return { contents: 'export default "test-file-stub";', loader: "js" };
-      }
-
-      if (args.path === "virtual:web-test-svg-stub") {
-        return {
-          contents: `
-import { createElement, forwardRef } from "react";
-const SvgrMock = forwardRef((props, ref) => createElement("span", { ref, ...props }));
-SvgrMock.displayName = "SvgrMock";
-export const ReactComponent = SvgrMock;
-export default SvgrMock;
-`,
-          loader: "tsx",
-        };
-      }
-    });
-  },
-});
-
-const dom = new JSDOM("<!DOCTYPE html><html><body></body></html>", {
-  pretendToBeVisual: true,
-  url: "http://localhost/",
-});
-
-const { window } = dom;
-
-globalThis.window = window as unknown as Window & typeof globalThis;
-globalThis.document = window.document;
-globalThis.navigator = window.navigator;
-globalThis.location = window.location;
-globalThis.history = window.history;
-globalThis.localStorage = window.localStorage;
-globalThis.sessionStorage = window.sessionStorage;
-globalThis.HTMLElement = window.HTMLElement;
-Object.defineProperty(window, "HTMLIFrameElement", {
-  configurable: true,
-  value: window.HTMLElement,
-  writable: true,
-});
-globalThis.HTMLIFrameElement = window.HTMLIFrameElement;
-globalThis.HTMLAnchorElement = window.HTMLAnchorElement;
-globalThis.Node = window.Node;
-globalThis.Event = window.Event;
-globalThis.CustomEvent = window.CustomEvent;
-globalThis.MouseEvent = window.MouseEvent;
-globalThis.KeyboardEvent = window.KeyboardEvent;
-// Bun's native globalThis.dispatchEvent/addEventListener operate on Bun's own
-// Event realm. Dexie constructs `new CustomEvent(...)` against the jsdom
-// Event class above, so dispatching through Bun's native EventTarget throws
-// "must be an instance of Event" (cross-realm brand check). Rebind these to
-// jsdom's window so the whole event pipeline stays in one realm.
-globalThis.dispatchEvent = window.dispatchEvent.bind(window);
-globalThis.addEventListener = window.addEventListener.bind(window);
-globalThis.removeEventListener = window.removeEventListener.bind(window);
-globalThis.self = window;
-globalThis.IS_REACT_ACT_ENVIRONMENT = true;
-
-const noopAlert = () => {};
-window.alert = noopAlert;
-globalThis.alert = noopAlert;
-
-class MockObserver<_T> implements IntersectionObserver, ResizeObserver {
-  root!: Document | Element | null;
-  rootMargin!: string;
-  thresholds!: readonly number[];
-  observe = (...args: unknown[]): unknown => args;
-  unobserve = (...args: unknown[]): unknown => args;
-  disconnect = (...args: unknown[]): unknown => args;
-  takeRecords = (): IntersectionObserverEntry[] => [];
-}
-
-class MediaQuery implements MediaQueryList {
-  matches = false;
-  media: string;
-  onchange = null;
-
-  constructor(query: string) {
-    this.media = query;
-  }
-
-  addListener(): void {}
-
-  removeListener(): void {}
-
-  removeEventListener(): void {}
-
-  addEventListener(): void {}
-
-  dispatchEvent(): boolean {
-    return true;
-  }
-}
-
-function getPointerEvent(mouseEvent: typeof globalThis.MouseEvent) {
-  class PointerEvent extends mouseEvent implements globalThis.PointerEvent {
-    altitudeAngle: number;
-    azimuthAngle: number;
-    height: number;
-    isPrimary: boolean;
-    pointerId: number;
-    pointerType: string;
-    pressure: number;
-    tangentialPressure: number;
-    tiltX: number;
-    tiltY: number;
-    twist: number;
-    width: number;
-
-    constructor(type: string, eventInitDict?: PointerEventInit) {
-      super(type, eventInitDict);
-      this.altitudeAngle = eventInitDict?.altitudeAngle ?? 0;
-      this.azimuthAngle = eventInitDict?.azimuthAngle ?? 0;
-      this.height = eventInitDict?.height ?? 1;
-      this.isPrimary = eventInitDict?.isPrimary ?? false;
-      this.pointerId = eventInitDict?.pointerId ?? 0;
-      this.pointerType = eventInitDict?.pointerType ?? "";
-      this.pressure = eventInitDict?.pressure ?? 0;
-      this.tangentialPressure = eventInitDict?.tangentialPressure ?? 0;
-      this.tiltX = eventInitDict?.tiltX ?? 0;
-      this.tiltY = eventInitDict?.tiltY ?? 0;
-      this.twist = eventInitDict?.twist ?? 0;
-      this.width = eventInitDict?.width ?? 1;
-    }
-
-    getCoalescedEvents(): globalThis.PointerEvent[] {
-      throw new Error("Method not implemented.");
-    }
-
-    getPredictedEvents(): globalThis.PointerEvent[] {
-      throw new Error("Method not implemented.");
-    }
-  }
-
-  return PointerEvent;
-}
-
-// Tailwind v4 emits valid browser CSS that jsdom 26 cannot parse. The web
-// tests assert DOM behavior and inline layout values, so injecting the full
-// generated stylesheet only adds parser noise without increasing coverage.
-
-window.HTMLElement.prototype.scroll = () => {};
-window.HTMLElement.prototype.scrollIntoView = () => {};
-window.scrollTo = () => {};
-window.document.elementFromPoint = () => null;
-window.PointerEvent = getPointerEvent(window.MouseEvent);
-
-window.fetch = globalThis.fetch.bind(globalThis);
-window.Blob = globalThis.Blob;
-window.File = globalThis.File;
-window.FormData = globalThis.FormData;
-window.Event = globalThis.Event;
-window.CustomEvent = globalThis.CustomEvent;
-window.MouseEvent = globalThis.MouseEvent;
-window.KeyboardEvent = globalThis.KeyboardEvent;
-window.Headers = globalThis.Headers;
-window.Request = globalThis.Request;
-window.Response = globalThis.Response;
-window.XMLHttpRequest = globalThis.XMLHttpRequest;
-window.ArrayBuffer = globalThis.ArrayBuffer;
-window.Uint8Array = globalThis.Uint8Array;
-window.Uint8ClampedArray = globalThis.Uint8ClampedArray;
-window.Uint16Array = globalThis.Uint16Array;
-window.Uint32Array = globalThis.Uint32Array;
-window.Int8Array = globalThis.Int8Array;
-window.Int16Array = globalThis.Int16Array;
-window.Int32Array = globalThis.Int32Array;
-window.Float32Array = globalThis.Float32Array;
-window.Float64Array = globalThis.Float64Array;
-window.DataView = globalThis.DataView;
-window.SharedArrayBuffer = globalThis.SharedArrayBuffer;
-window.Atomics = globalThis.Atomics;
-window.WebAssembly = globalThis.WebAssembly;
-
-window.URL.createObjectURL = globalThis.URL.createObjectURL.bind(window.URL);
-window.URL.revokeObjectURL = globalThis.URL.revokeObjectURL.bind(window.URL);
-
-window.IntersectionObserver =
-  MockObserver<IntersectionObserverCallback> as unknown as typeof IntersectionObserver;
-window.ResizeObserver =
-  MockObserver<ResizeObserverCallback> as unknown as typeof ResizeObserver;
-
-for (const key of Object.getOwnPropertyNames(window)) {
-  if (key in globalThis) {
-    continue;
-  }
-
-  const descriptor = Object.getOwnPropertyDescriptor(window, key);
-
-  if (descriptor) {
-    try {
-      Object.defineProperty(globalThis, key, descriptor);
-    } catch {
-      // Some properties are non-configurable in Bun's globalThis; skip them.
-    }
-  }
-}
-
-Object.defineProperty(window, "matchMedia", {
-  writable: true,
-  value: (query: string): MediaQueryList => new MediaQuery(query),
-});
-
-const originalGetContext = window.HTMLCanvasElement.prototype.getContext;
-window.HTMLCanvasElement.prototype.getContext = function (
-  this: HTMLCanvasElement,
-  contextId: string,
-  ...args: unknown[]
-) {
-  if (contextId === "2d") {
-    return {
-      font: "",
-      measureText: (text: string) => ({ width: text.length * 7 }),
-    } as unknown as CanvasRenderingContext2D;
-  }
-
-  return originalGetContext.apply(this, [contextId, ...args] as never);
-};
-
-await import("fake-indexeddb/auto");
-// fake-indexeddb/auto attaches indexedDB (and friends) to `typeof window
-// !== "undefined" ? window : ...`, which resolves to the jsdom window set
-// above. Dexie reads `globalThis.indexedDB`/`globalThis.IDBKeyRange`
-// directly, so mirror them onto globalThis or `new Dexie(...).open()`
-// throws MissingAPIError.
-globalThis.indexedDB = window.indexedDB;
-globalThis.IDBKeyRange = window.IDBKeyRange;
-globalThis.IDBCursor = window.IDBCursor;
-globalThis.IDBCursorWithValue = window.IDBCursorWithValue;
-globalThis.IDBDatabase = window.IDBDatabase;
-globalThis.IDBFactory = window.IDBFactory;
-globalThis.IDBIndex = window.IDBIndex;
-globalThis.IDBObjectStore = window.IDBObjectStore;
-globalThis.IDBOpenDBRequest = window.IDBOpenDBRequest;
-globalThis.IDBRequest = window.IDBRequest;
-globalThis.IDBTransaction = window.IDBTransaction;
-globalThis.IDBVersionChangeEvent = window.IDBVersionChangeEvent;
-
-if (typeof globalThis.structuredClone === "undefined") {
-  globalThis.structuredClone = (obj: unknown) => {
-    return JSON.parse(JSON.stringify(obj));
-  };
-}
 
 mockNodeModules();
 
-const sessionModule = await import("supertokens-web-js/recipe/session");
-const { cleanup, configure } = await import("@testing-library/react");
-const { resetAllStores } = await import("./utils/state/reset-stores");
-
-// Give async queries (findBy*, waitFor) real headroom. The default 1000ms is
-// fine locally (the whole suite runs in ~16s) but too tight on CI, where every
-// matrix job shares the runner's cores: under that contention a component that
-// resolves in tens of ms locally can take past a second, timing out queries in
-// tests that are otherwise correct (AuthModal/SelectView flaked this way). A
-// genuinely stuck async never resolves regardless, so this only widens the
-// window for slow-but-correct ones — it cannot mask a real hang.
-configure({ asyncUtilTimeout: 5000 });
-
-function resetDocument() {
-  document.body.innerHTML = "";
-  document.body.removeAttribute("style");
-  document.body.removeAttribute("class");
-  document.body.removeAttribute("data-app-locked");
-  document.documentElement.removeAttribute("style");
-  for (const style of document.head.querySelectorAll("style")) {
-    if (
-      !style.textContent?.includes(":has(.react-datepicker__day--selected)")
-    ) {
-      continue;
-    }
-
-    style.textContent = style.textContent.replaceAll(
-      /[^{}]+:has\(\.react-datepicker__day--selected\)[^{]*\{[^{}]*\}/g,
-      "",
-    );
-  }
-}
-
-function resetBrowserState() {
-  dom.reconfigure({ url: "http://localhost/" });
-  localStorage.clear();
-  sessionStorage.clear();
-}
-
-beforeEach(() => {
-  sessionModule.doesSessionExist?.mockResolvedValue(true);
-});
-
-beforeAll(() => server.listen({ onUnhandledRequest: "error" }));
-afterEach(async () => {
-  await Promise.resolve();
-  cleanup();
-  resetDocument();
-  resetBrowserState();
-  resetAllStores();
-  server.resetHandlers();
-});
-// Safety net when multiple files share a --parallel worker: the last test in a
-// file runs afterEach above; this mirrors that reset if Bun ever skips it.
-afterAll(() => {
-  resetDocument();
-  resetBrowserState();
-  resetAllStores();
-});
-afterAll(() => server.close());
-afterAll(() => mock.restore());
+// Dynamic import keeps lifecycle hooks (and session import) after mock.module
+// registration — static imports are hoisted and would run mockNodeModules too late.
+await import("./setup/test-lifecycle");
