+++ tsconfig.json
@@ -16,7 +16,8 @@
     "**/*.spec.ts",
     "**/*.spec.tsx",
     "packages/scripts/src/testing/core.preload.ts",
-    "packages/scripts/src/testing/scripts.preload.ts"
+    "packages/scripts/src/testing/scripts.preload.ts",
+    "packages/scripts/src/testing/scripts.preload.fast.ts"
   ],
   "compilerOptions": {
     "target": "ES2022" /* Specify ECMAScript target version. */,
